Added lookup and update for SPSegment and some functions of SlottedPage

+++ include/slottedpages/tid.h
@@ -10,13 +10,13 @@
 class TID {
 private:
     uint64_t page_id;
-    uint32_t slot_id;
+    uint16_t slot_id;
 public:
-    TID(uint64_t page_id, uint32_t slot_id) : (this->page_id, this->slot_id);
+    TID(uint64_t page_id, uint16_t slot_id) : (this->page_id, this->slot_id);
 
     uint64_t getPageId() { return page_id; };
 
-    uint32_t getSlotId() { return slot_id; };
+    uint16_t getSlotId() { return slot_id; };
 };
 
 #endif //DBLITE_TID_H
